Explicitly added a way to configure a Jpeg editor, for simplicity of use.

+++ src/labs_coordinate_pictures/TestsSortFiles.cs
@@ -150,7 +150,11 @@ namespace labs_coordinate_pictures
             try
             {
                 string dir = TestUtil.GetTestWriteDirectory();
-                Directory.Delete(dir, true);
+
+                if (Directory.Exists(dir))
+                {
+                    Directory.Delete(dir, true);
+                }
             }
             catch (Exception)
             {
